==> notebooks/webscrape/04 - Validating Web Scraped Data - Episodes.ipynb <==
import pandas as pd
import json
import os
import sys

# import parser
from src.scraping.parse_chapter import parse_chapter

# remove column width restrictions
pd.set_option('display.max_colwidth', None)

ROOT = os.path.abspath(os.path.join(os.path.dirname(os.getcwd()), '..'))
data_path = os.path.join(ROOT, 'data', 'raw')
one_piece_episodes_path = os.path.join(data_path, 'one_piece_episodes.json')

df = pd.read_json(one_piece_episodes_path)

df.info()

# print url where short summary is NaN
df[df['short_summary'].isna()]['url']

# print url where long summary is NaN
df[df['long_summary'].isna()]['url']

# print url where characters is nan
# this episode hasn't aired yet
df[df['characters'].isna()]['url']
