==> notebooks/webscrape/07 - Validate Web Scraped Data - Characters.ipynb <==
import pandas as pd
import json
import os
import sys


# remove column width restrictions
pd.set_option('display.max_colwidth', None)

ROOT = os.path.abspath(os.path.join(os.path.dirname(os.getcwd()), '..'))
data_path = os.path.join(ROOT, 'data', 'raw')
one_piece_characters_path = os.path.join(data_path, 'one_piece_characters_data.jsonl')

df = pd.read_json(one_piece_characters_path, lines=True)

df.isna().sum()

df.info()

## See random sample of no affiliations

no_affiliations_sample = df[df['affiliations'].isna()][['name','url']].sample(5)

for index, row in no_affiliations_sample.iterrows():
    print(f"Name: {row['name']}, URL: {row['url']}")

df[~df['devil_fruit'].isna()][['name','url','devil_fruit']]

#### Bounty

# check with bounty
df[~df['bounty'].isna()][['name','url','bounty']]

# get bounties that are less than 5 digits
# these are mostly incorrectly parsed bounties
len(df['bounty'].dropna()[df['bounty'].dropna().astype(str).str.len() < 5])

# read the file again and then set the bounty to none if it is less than 5 digits
df = pd.read_json(one_piece_characters_path, lines=True)
df.loc[df['bounty'].dropna()[df['bounty'].dropna().astype(str).str.len() < 5].index, 'bounty'] = None

# save back to jsonl
# df.to_json(one_piece_characters_path, orient='records', lines=True)

### Check anime debut is null

# anime debut is null
print(f"Number of characters with null anime debut: {len(df[df['anime_debut'].isna()])}")
df[df['anime_debut'].isna()][['name','url']].head(10)

- The above links have "first appearance" as keyword instead of debut. 
- Formatting issues

df.info()

### Check missing general info

### Check missing general info

df[df['general_info'].isna()][['name','url']]

- Most likely bug because it points to Ussop

#### Check no appearance

# no appearance
df[df['appearance'].isna()][['name','url']]

- Most of links above or not characters but rather groups. 
- Could be because link is available but not characters really
